requirements.txt
requests~=2.32.0
pytest~=8.3.4
pyodbc~=5.2.0
pandas~=2.2.3
pytz~=2024.2
PyYAML~=6.0.2